+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Pushover, Telegram and LaMetric, contributed by [@lollox80](https://github.com/lollox80)
 ### Multimedia
 - New `media_url_prefix` in platform settings to control the URL path which Home Assistant uses to expose the media directory. Make this empty to stop the directory being available via HA web UI
+- Fix media cleanup job handling of files in subdirectories
 ### Mobile Push
 - By default, thumbnail image is captured at time of notification
 - Better Android integration